local/requirements.txt
gymnasium[mujoco]>=1.0.0
stable-baselines3>=2.3.0
tensorboard>=2.15.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
